pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-mapreduce"
version = "0.1.0"
description = "Scalable Retrieval-Augmented Generation with sequential and Ray MapReduce pipelines."
readme = "README.md"
authors = [{name = "COMP4651 Team"}]
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-text-splitters>=0.2.0",
    "langchain-ollama>=0.1.0",
    "ray[default]>=2.10.0",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.7.0",
    "typer>=0.12.3",
    "rich>=13.7.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "ruff>=0.5.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
